nRF24: Avoid channel switch on Thing
Removes channel-switching behaviour on thing. If it doesn't have
a connection, it will reamin on MGMT channel until a connection
is made. If it is already connected, thing remains on RAW while
the connection exists and only move back to MGMT in case the
connection is lost (e.g. timeout).